add function to run forest models to util.py
